refactor(BridgeGame): BridgeGame 파일 Controller 디렉토리로 이동
